+++ lib/services_worker.dart
@@ -1,8 +1,8 @@
 /// A lightweight Dart utility for safely running synchronous,
-/// asynchronous, and isolate-based tasks.
+/// asynchronous, and isolate-based tasks with structured responses.
 library;
 
-export 'src/services_error.dart';
 export 'src/services_exception.dart';
+export 'src/services_failure.dart';
 export 'src/services_response.dart';
 export 'src/services_worker.dart';